workoutplanner/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
boto3==1.17.1
botocore==1.20.1
cement==2.8.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.1.5
django-countries==7.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
filelock==3.0.12
future==0.16.0
idna==2.10
jmespath==0.10.0
Markdown==3.3.3
pathspec==0.5.9
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.4
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
yapf==0.30.0
